requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
psycopg[binary,pool]==3.2.1
httpx==0.27.2
python-dotenv==1.0.1

